# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Deligo Mapper API",
    "version": "1.0.0",
    "description": "Delivery mapping and tracking service for Deligo. Handles delivery order lifecycle, driver location tracking, and address geocoding."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.deligo.mn",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key required for protected endpoints"
      }
    },
    "schemas": {
...
      "DriverLocation": {
        "type": "object",
        "required": ["latitude", "longitude"],
        "properties": {
          "latitude": { "type": "number", "format": "float", "example": 47.9184 },
          "longitude": { "type": "number", "format": "float", "example": 106.9177 }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Invalid or missing API key",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "detail": "Invalid or missing API key" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "detail": "Order not found" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/drivers/{driver_id}/location": {
      "post": {
        "tags": ["drivers"],
        "summary": "Update Driver Location",
        "description": "Driver reports their current GPS position. Creates or updates the driver location record.",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "1773113257387725" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DriverLocation" },
              "example": { "latitude": 47.9184, "longitude": 106.9177 }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Driver location updated",
            "content": {
              "application/json": {
                "example": {
                  "status": "ok",
                  "data": {
                    "driver_id": "1773113257387725",
                    "latitude": 47.9184,
                    "longitude": 106.9177,
                    "updated_at": "2026-04-14T10:00:00"
                  }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "get": {
        "tags": ["drivers"],
        "summary": "Get Driver Location",
        "description": "Get a driver's last known GPS position.",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "1773113257387725" }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver location",
            "content": {
              "application/json": {
                "example": {
                  "status": "ok",
                  "data": {
                    "driver_id": "1773113257387725",
                    "latitude": 47.9184,
                    "longitude": 106.9177,
                    "updated_at": "2026-04-14T10:00:00"
                  }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  }
}